added splash screen. fixed draft and other bugs. added check at generate presciption. fixed progress dialog. added storage permission check

+++ app/src/main/java/com/example/mobihealthapis/Models/Vitals.java
@@ -40,7 +40,7 @@ public class Vitals {
 
     public static class Data {
         @SerializedName("BMI")
-        private float BMI;
+        private double BMI;
         @SerializedName("Weight")
         private double Weight;
         @SerializedName("Height")
@@ -70,11 +70,11 @@ public class Vitals {
             Temperature = temperature;
         }
 
-        public float getBMI() {
+        public double getBMI() {
             return BMI;
         }
 
-        public void setBMI(float BMI) {
+        public void setBMI(double BMI) {
             this.BMI = BMI;
         }
